Refactor DocumentConstants to model package
- Moved `DocumentConstants.java` from `org.waveprotocol.box.common` to `org.waveprotocol.wave.model.document` to align with the codebase's model packages.
- Addressed the TODO in the original file.
- Updated all references across the project.
